Integrate genshi.scss in the docs site
Given that Astro can integrate with SASS and compile in the process
.scss files, this commit will replace the `<link>` tag that would pick
Genshi's bundle with an import statement that just imports genshi.scss.
Astro will take care of compiling the stylesheet and hot reloading it
during development. On deployment, it will bundle the stylesheet.

I've changed the run scripts of this package, because if Astro can deal
with the stylesheet, there is a lot of things that are not required
anymore. Specifically, the steps that manually copy files from dist/ to
www/ are now gone, so the task docs:copy is gone.

Some tasks have also been renamed:

- build is now known as dist, and it compiles Genshi into genshi.css.
  When a bundler is going to use this file, this is the suggested
  artifact anyway.
- docs:build is now just docs, because everything can be done in a
  single step to compile the docs.

The new task dist:min will compile a minified version of genshi.scss
into genshi.min.css, in case manual import wants to be done anyway over
a CDN.

Dependencies that are not in use anymore have been removed from the
project, such as copyfiles and npm-run-all.

References:
- https://docs.astro.build/en/guides/styling/#import-a-local-stylesheet
- https://docs.astro.build/en/guides/styling/#sass-and-scss