Optimize logo, add dark variant
Given that the logo is just a bunch of <text> nodes, this commit
replaces the SVG version of the logo with a hand-made version that is
specifically designed to be precise, and also minimal.

The color scheme of the logo is changed, and also now two versions of
the logo are rendered in light and dark mode. The layout and the README
files have been modified to use the dark variant when available.