Force Col and Stack to show gap in mobile mode
It is important that the elements that are part of a col or a stack keep
the gap even when stacked vertically on mobile. Otherwise, the elements
will look too close.

This is interesting because one might think that the flex mode is
enabled after the breakpoint triggers, but actually the flex mode should
always be enabled, and just switch to a horizontal layout once there is
enough space.