Convert the color palettes back to hex
The colors.mix() function is not generating hexadecimal colors. Instead,
it is generating rgb() colors with too many decimals. We don't need so
many decimals, so this commit adds a function that converts back to
hexadecimal.

This will save size in the compiled file. The browser may have a better
time decoding these new colors, but I doubt that it was an issue in
first place.